requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
beautifulsoup4==4.6.0
billiard
bleach==2.0.0
celery==3.1.25
colorama==0.3.9
contextlib2==0.5.5
diff-match-patch==20110725.1
Django==1.6.11
django-braces==1.8.0
django-celery==3.0.23
django-debug-toolbar==1.2.1
django-filter==0.11.0
django-haystack==2.4.0
django-mailgun==0.2.2
django-model-utils==2.2
django-redis==4.2.0
-e git+https://github.com/guoku/django-sendcloud.git@6a8cee55f03ee5a7c819b15901969e40edc7916d#egg=django_sendcloud
django-simple-captcha==0.4.5
-e git+https://github.com/guoku/django-storages.git@551f871ab11fc7261e63d97c2c6b02f5f8922e16#egg=django_storages
django-uuslug==1.1.8
djangorestframework==3.0.0
ecdsa==0.11
factory-boy==2.6.0
fake-factory==9999.9.9
funcsigs==1.0.2
gevent==1.2.2
greenlet==0.4.12
gunicorn==18.0
hiredis==0.2.0
html5lib==0.999999999
jieba==0.35
kombu==3.0.37
librabbitmq==1.6.1
marshmallow==2.9.1
mock==1.3.0
msgpack-python==0.4.8
MySQL-python==1.2.5
packaging==16.8
paramiko==1.14.1
pbr==3.1.1
Pillow==2.8.2
py==1.4.34
pycrypto==2.6.1
pymogile==2.0.2
pyparsing==2.2.0
pysolr==3.3.2
pytest==2.7.2
pytest-django==2.9.1
python-dateutil==2.3
python-slugify==1.2.4
pytz==2014.10
qrcode==5.2.2
raven==6.1.0
redis==2.10.3
requests==2.6.0
sinaweibopy==1.1.3
six==1.10.0
sqlparse==0.1.11
tablib==0.10.0
unicodecsv==0.14.0
Unidecode==0.4.21
Wand==0.4.0
webencodings==0.5.1
xpinyin==0.5.3
django-environ==0.4.4
oss2==2.4.0
arrow==0.12.0